+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## v0.1.1
 
 - Deck
-  - Feat: save config to file. [#2](https://github.com/DiscreteTom/stickdeck-rs/issues/2).
+  - Feat: save config to file. [#2](https://github.com/DiscreteTom/stickdeck-rs/issues/2)
+  - Feat: add dark mode. [#1](https://github.com/DiscreteTom/stickdeck-rs/issues/1)
   - Note: change default input update interval to 3ms to reach the max update rate (about 250 updates per second).
 
 ## v0.1.0