+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 - Client (PC)
   - Note: apply `env_logger`, add debug and trace log.
-- Server (SteamDeck)
+- Server (Steam Deck)
   - Feat: save config to file. [#2](https://github.com/DiscreteTom/stickdeck-rs/issues/2)
   - Feat: add dark mode. [#1](https://github.com/DiscreteTom/stickdeck-rs/issues/1)
   - Note: change default input update interval to 3ms to reach the max update rate (250+ updates per second).